==> notebooks/4_train_args_1.ipynb <==
## Setting up Training parameters and Hyperparameters

If you have selected this notebook then it is assumed that you have selected to go ahead with SGD (Stochastic Gradient Descent) as the optimizer.

### Training Parameters and Hyperparameters Explanation

1.  **`output_dir`**: Specifies the directory where model checkpoints and training logs will be saved.

2.  **`evaluation_strategy`**: Defines the strategy for evaluating the model during training.

3.  **`save_strategy`**: Specifies when to save model checkpoints.

4.  **`learning_rate`**: Determines the step size at which the optimizer adjusts model weights during training.

5.  **`per_device_train_batch_size`**: Specifies the batch size for training data per GPU, impacting memory usage and computational efficiency.

6.  **`per_device_eval_batch_size`**: Sets the batch size for evaluation data per GPU, affecting memory and computation during evaluation.

7.  **`num_train_epochs`**: Indicates the total number of training epochs, which are complete passes through the training dataset.

8.  **`warmup_ratio`**: Determines the ratio of warmup steps to the total number of training steps, helping the optimizer to smoothly adapt in the initial stages of training.

9.  **`weight_decay`**: Introduces L2 regularization to the optimizer, helping to prevent overfitting by penalizing large model weights.

10. **`load_best_model_at_end`**: Specifies whether to load the best model based on the chosen evaluation metric at the end of training.

11. **`metric_for_best_model`**: Specifies the evaluation metric used to determine the best model, which is set to "accuracy" in this case.

12. **`save_total_limit`**: Sets the maximum number of model checkpoints to keep, preventing excessive storage usage.

13. **`logging_dir`**: Designates the directory where training logs, such as training progress and performance metrics, will be stored.

14. **`optimizers`**: Specifies the optimizers used for model parameter updates during training, allowing for gradient-based optimization algorithms like AdamW, SGD and many more.

## Implementation

### Importing relevent Library

import json


### Setting up the training arguments

args = {
    "output_dir": "./output",
    "evaluation_strategy": "epoch",
    "save_strategy": "epoch",
    "per_device_train_batch_size": 8,
    "per_device_eval_batch_size": 8,
    "num_train_epochs": 10,
    "warmup_ratio": 0.1,
    "weight_decay": 0.001,
    "load_best_model_at_end": True,
    "metric_for_best_model": "accuracy",
    "save_total_limit": 1,
    "logging_dir": "./logs",
    "optimizer": "sgd"
}



### Storing the dictionary to a json file


with open('args.json', 'w') as args_file:
    json.dump(args, args_file, indent=4)


## Next Steps:

Now we have completed all the steps needed before training a model, The next step is Training the model and obtaining the final result. But wait, the author talked in the paper in Training and evaluation section that they did hyperparameter tuning but they never mentioned that they did it on all three models or just the largest model. Again why this doubt comes because if they are comparing then are they comparing it on the same ground or the models specific performance ?

Let’s understand this in details. In our scenario, we are training three models 1) model trained on largest dataset. 2) model trained on full few shot data 3) model trained on full few shot data and it’s augmented version. We are performing hyper parameter search on Learning rate as the author mentioned that they performed search on 4 standard learning rates used for training bert classifier. Now we are confused if we need to perform search on the standard dataset and then use the same on the other two or we are performing seperately on different datasets.

So inshort we are left with two choices :

-   Hyperparameter tuning on the largest model

-   Hyperparameter tuining on all three models

It’s upto you that what you are picking, below we have two notebook each with a different choice. Pick your own adventure and see how similar are your results.

-   [Notebook(Hyperparameter tuning on the largest model)](./5_model_training_1.ipynb)

-   [Notebook(Hyperparameter tuining on all three models)](./5_model_training_2.ipynb)
